--- app/src/main/java/com/example/appdevelopment/data/MyViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.appdevelopment.data
-
-import androidx.lifecycle.ViewModel
-import com.example.appdevelopment.data.domain.repository.MyRepository
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class MyViewModel @Inject constructor(
-    private val repository: MyRepository
-): ViewModel() {
-
-}
