+++ app/src/main/java/com/example/appdevelopment/data/di/AppModule.kt
@@ -1,41 +1,23 @@
 package com.example.appdevelopment.data.di
 
-import com.example.appdevelopment.data.AuthRepository
-import com.example.appdevelopment.data.AuthRepositoryImpl
-import com.example.appdevelopment.data.domain.repository.MyRepository
-import com.example.appdevelopment.data.remote.MyApi
-import com.example.appdevelopment.data.remote.repository.MyRepositoryImpl
+import com.example.appdevelopment.data.domain.repository.AuthRepository
+import com.example.appdevelopment.data.remote.repository.AuthRepositoryImpl
 import com.google.firebase.auth.FirebaseAuth
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import retrofit2.Retrofit
-import retrofit2.create
 import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 object AppModule {
-
     @Provides
+    @Singleton
     fun provideFirebaseAuth(): FirebaseAuth = FirebaseAuth.getInstance()
 
     @Provides
+    @Singleton
     fun provideAuthRepository(impl: AuthRepositoryImpl): AuthRepository = impl
 
-    @Provides
-    @Singleton
-    fun provideMyApi(): MyApi {
-        return Retrofit.Builder()
-            .baseUrl("https://test.com")
-            .build()
-            .create(MyApi::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideMyRepository(api: MyApi): MyRepository{
-        return MyRepositoryImpl(api)
-    }
 }
