--- app/src/main/java/com/example/appdevelopment/data/domain/repository/MyRepository.kt
@@ -1,6 +1,0 @@
-package com.example.appdevelopment.data.domain.repository
-
-interface MyRepository {
-
-    suspend fun networkCall()
-}
