--- app/src/main/java/com/example/appdevelopment/data/remote/MyApi.kt
@@ -1,9 +1,0 @@
-package com.example.appdevelopment.data.remote
-
-import retrofit2.http.GET
-
-interface MyApi {
-
-    @GET("test")
-    suspend fun networkCall()
-}
