--- app/src/main/java/com/example/appdevelopment/data/remote/repository/MyRepositoryImpl.kt
@@ -1,13 +1,0 @@
-package com.example.appdevelopment.data.remote.repository
-
-import com.example.appdevelopment.data.domain.repository.MyRepository
-import com.example.appdevelopment.data.remote.MyApi
-
-class MyRepositoryImpl(
-    private val api: MyApi
-): MyRepository {
-
-    override suspend fun networkCall() {
-
-    }
-}
