+++ app/src/main/java/com/example/appdevelopment/ui/screens/loginView/LoginEvent.kt
@@ -3,4 +3,5 @@ package com.example.appdevelopment.ui.screens.loginView
 sealed class LoginEvent {
     data class OnEmailChanged(val email: String): LoginEvent()
     data class OnPasswordChanged(val password: String): LoginEvent()
+    object OnLogin: LoginEvent()
 }
