+++ app/src/main/java/com/example/appdevelopment/ui/screens/loginView/LoginViewModel.kt
@@ -3,9 +3,9 @@ package com.example.appdevelopment.ui.screens.loginView
 import androidx.compose.runtime.collectAsState
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.appdevelopment.data.AuthRepository
 import com.example.appdevelopment.data.AuthViewModel
 import com.example.appdevelopment.data.Resource
+import com.example.appdevelopment.data.domain.repository.AuthRepository
 import com.google.firebase.auth.FirebaseUser
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -16,16 +16,17 @@ import javax.inject.Inject
 
 @HiltViewModel
 class LoginViewModel @Inject constructor(
-    private val viewModel: AuthViewModel?
+    private val repository: AuthRepository
 ) : ViewModel() {
 
-    val loginFlow = viewModel?.loginFlow?.collectAsState()
+    private val _loginFlow = MutableStateFlow<Resource<FirebaseUser>?>(null)
+    val loginFlow: StateFlow<Resource<FirebaseUser>?> = _loginFlow
 
     private val _uiState = MutableStateFlow(LoginUIState())
     val uiState = _uiState.asStateFlow()
 
     fun onEvent(event: LoginEvent){
-        loginFlow?.value?.let {
+        /*loginFlow?.value?.let {
             when(it){
                 is Resource.Failure -> TODO()
                 Resource.Loading -> TODO()
@@ -33,10 +34,11 @@ class LoginViewModel @Inject constructor(
                     TODO()
                 }
             }
-        }
+        }*/
         when(event){
             is LoginEvent.OnEmailChanged -> onEmailChanged(event.email)
             is LoginEvent.OnPasswordChanged -> onPasswordChanged(event.password)
+            is LoginEvent.OnLogin -> login(_uiState.value.emailText, _uiState.value.passwordText)
         }
     }
 
@@ -46,6 +48,18 @@ class LoginViewModel @Inject constructor(
 
     fun onPasswordChanged(password: String){
         _uiState.value = _uiState.value.copy(passwordText = password)
+    }
+
+    fun login(email: String, password: String) = viewModelScope.launch {
+        _loginFlow.value = Resource.Loading
+        val result = repository.login(email, password)
+        _loginFlow.value = result
+    }
+
+    init {
+        if(repository.currentUser != null){
+            _loginFlow.value = Resource.Success(repository.currentUser!!)
+        }
     }
